xdtools: add critical security check, just in case
If (whyever, however) someone deletes the build directory, 'gettop' does not work correctly anymore.

If you try to cd to $(gettop), it'll execute "cd ", what changes the directory to the home directory.

By running a command like 'reporesync full' you run into big trouble. It'll remove the whole home directory.

To prevent that, I have added a special security check which aborts the process if you are currently on the home directory.

Additionally, a prompt has been added to clarify what you are really doing and to confirm the operation.

PLEASE BE CAREFUL!!